Refactored jarbin stuff to be called what I think it should have been called and thats varbin. Jacob should look over this and make sure that looks OK. Also modified VariantPool to check for a IonTorrent attribute at initialization of the xml object which will create a special type of Ion Torrent CSVLineReader instead of the traditional one otherwise.
